+++ src/main/java/org/cotato/backend/recruit/admin/service/staffEvaluation/StaffEvaluationService.java
@@ -3,7 +3,7 @@ package org.cotato.backend.recruit.admin.service.staffEvaluation;
 import lombok.RequiredArgsConstructor;
 import org.cotato.backend.recruit.admin.dto.request.staffEvaluation.CreateStaffEvaluationRequest;
 import org.cotato.backend.recruit.admin.dto.response.staffEvaluation.StaffEvaluationResponse;
-import org.cotato.backend.recruit.admin.service.applicationAdmin.ApplicationAdminService;
+import org.cotato.backend.recruit.admin.service.application.ApplicationAdminService;
 import org.cotato.backend.recruit.domain.application.entity.Application;
 import org.cotato.backend.recruit.domain.evaluation.entity.Evaluation;
 import org.cotato.backend.recruit.domain.evaluation.enums.EvaluatorType;
@@ -21,9 +21,10 @@ public class StaffEvaluationService {
 
 	public StaffEvaluationResponse getEvaluation(Long applicationId, EvaluatorType evaluatorType) {
 		Application application = applicationAdminService.getApplication(applicationId);
-		Evaluation evaluation = evaluationRepository
-				.findByApplicationAndEvaluatorType(application, evaluatorType)
-				.orElse(null);
+		Evaluation evaluation =
+				evaluationRepository
+						.findByApplicationAndEvaluatorType(application, evaluatorType)
+						.orElse(null);
 
 		return StaffEvaluationResponse.from(evaluation != null ? evaluation.getComment() : null);
 	}
@@ -32,9 +33,10 @@ public class StaffEvaluationService {
 	public void createEvaluation(Long applicationId, CreateStaffEvaluationRequest request) {
 		Application application = applicationAdminService.getApplication(applicationId);
 
-		Evaluation evaluation = evaluationRepository
-				.findByApplicationAndEvaluatorType(application, request.evaluatorType())
-				.orElse(null);
+		Evaluation evaluation =
+				evaluationRepository
+						.findByApplicationAndEvaluatorType(application, request.evaluatorType())
+						.orElse(null);
 
 		if (evaluation != null) {
 			update(evaluation, request);
